requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
